API 15

+++ PartySortPlus/PartySortPlus.cs
@@ -87,7 +87,7 @@ public unsafe class PartySortPlus: IDalamudPlugin
         C = null;
     }
 
-    private void TerritoryChanged(ushort id)
+    private void TerritoryChanged(uint id)
     {
         SoftForceUpdate = true;
     }
